Read only necessary sensor data from VL53L5CX to reduce reading time. Read from I2C device directly because the API breaks when the intended #defines are made to disable reading of unnecessary data

+++ src/Sensor.h
@@ -10,6 +10,6 @@ constexpr uint8_t SENSOR_IMAGE_SIZE = SENSOR_IMAGE_WIDTH * SENSOR_IMAGE_WIDTH;
 constexpr int PIN_POWER = 2;
 
 int8_t sensor_setup();
-int8_t sensor_read(float distances[], uint16_t sigma[]);
+int8_t sensor_read(uint16_t distances[], uint16_t sigma[]);
 
 #endif
